fix(engine): base decode and prefill on what a token actually reads
Four fixes from Codex review on PR #1. They interleave in types.ts, speed.ts
and build-catalog.ts, so they land together rather than as commits that would
not build individually.

1. INT8 gets its own computeDtype. `flops.int8` was catalogued but unreachable:
   the int8 quant declared `fp8`, and peakFlops checked only `f.fp8`. The 3090
   and 4090 publish int8 and no fp8, so both silently prefilled at the fp16
   rate — measured 2.00x under. The reverse fallback is deliberately one-way:
   Ampere has INT8 tensor cores and no FP8 units, so lending an FP8 quant the
   INT8 rate would invent a number that hardware cannot reach.

2. Decode stops charging weights it never reads. `activeWeightBytes` started
   from `totalParams - expertParams`, the residency figure, which re-adds the
   input embedding table and Gemma 3's vision tower. New `activeDenseParams`
   subtracts both — but the embedding only when it is *untied*, since a tied
   table is a full vocab matmul on every step. gpt-oss-20b decode was 31% low.

   Tied-ness is read from the absence of an `lm_head.weight` tensor, not from
   `config.tie_word_embeddings`: that key is undefined on both Gemma 3 repos
   even though they are tied, so trusting it would drop a 1.0B table that
   decode does read.

3. Prefill stops charging the output projection per prompt token. Logits are
   computed for the positions that need them — one, for generation — not for
   every token; llama.cpp gathers output rows before the LM head and vLLM
   slices hidden states before it. Was 16% high on gpt-oss-20b, 9% on Gemma 3.

4. Any seed failure now blocks the catalog write. The old threshold tolerated
   5 of 17 failures and still overwrote the committed artifact with exit 0,
   deleting 29% of the catalog on a transient HF error. `--allow-partial`
   keeps the escape hatch for a permanently gated repo.

Also fixes the expertFraction fallback, which mixed an embedding-excluded
numerator with an embedding-included subtrahend and implied 9.5% of gpt-oss-20b
experts per token where its config says 4 of 32.

The calibration constants were deliberately not retuned. Removing these errors
moved the Spark anchors from ~10% and ~6% under to ~19% and ~10% over, while
EPYC stayed within 1% — the old fit was partly absorbing the bugs. Re-centring
a fudge factor right after removing what it was masking is how the next error
gets hidden; all three anchors remain inside the +/-30% band.

Tests 131 -> 148, including guards for the tied path through the engine rather
than only as catalog data, and for the int8 lookup missing in both directions.

+++ src/data/quants.ts
@@ -40,8 +40,9 @@ export const QUANTS: readonly QuantSpec[] = [
     id: 'int8',
     label: 'INT8',
     bpw: 8,
-    // INT8 tensor cores run at roughly the FP8 rate on hardware that has both.
-    computeDtype: 'fp8',
+    // Its own dtype, not an alias for fp8: Ampere reaches 2x fp16 on INT8 while having no FP8
+    // tensor cores whatsoever, so the two rates have to be looked up separately.
+    computeDtype: 'int8',
     qualityNote: 'Near-lossless with good calibration.',
     source: 'https://arxiv.org/abs/2208.07339',
   },
